# Conflicts:
#   backend-spring/src/main/java/com/ssafy/pcgg/domain/recommend/entity/QuoteCandidateEntity.java
#   backend-spring/src/main/java/com/ssafy/pcgg/domain/recommend/repository/ChassisRepository.java
#   backend-spring/src/main/java/com/ssafy/pcgg/domain/recommend/repository/CoolerRepository.java
#   backend-spring/src/main/java/com/ssafy/pcgg/domain/recommend/repository/CpuRepository.java
#   backend-spring/src/main/java/com/ssafy/pcgg/domain/recommend/repository/GpuRepository.java
#   backend-spring/src/main/java/com/ssafy/pcgg/domain/recommend/repository/MainboardRepository.java
#   backend-spring/src/main/java/com/ssafy/pcgg/domain/recommend/repository/PowerRepository.java
#   backend-spring/src/main/java/com/ssafy/pcgg/domain/recommend/repository/QuoteRepository.java
#   backend-spring/src/main/java/com/ssafy/pcgg/domain/recommend/repository/RamRepository.java
#   backend-spring/src/main/java/com/ssafy/pcgg/domain/recommend/repository/SsdRepository.java

+++ backend-spring/src/main/java/com/ssafy/pcgg/domain/recommend/entity/ChassisEntity.java
@@ -6,14 +6,13 @@ import org.hibernate.annotations.CreationTimestamp;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
-import java.util.List;
 
 @Entity
 @Table(name = "part_case")
 public class ChassisEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Null
     @Column(length = 100)
