+++ backend-spring/src/main/java/com/ssafy/pcgg/domain/recommend/entity/PowerEntity.java
@@ -7,14 +7,13 @@ import org.hibernate.annotations.CreationTimestamp;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
-import java.util.List;
 
 @Entity
 @Table(name = "part_power")
 public class PowerEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(length = 50)
     private String name;
